% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw.R
\name{dtw}
\alias{dtw}
\title{Performs dynamic time warping}
\usage{
dtw(time_series_1, time_series_2)
}
\arguments{
\item{time_series_1}{first time series}

\item{time_series_2}{second time series}
}
\value{
gods now what
}
\description{
Performs dynamic time warping for two input time series.
}
